backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.8",
    "greenlet>=3.1.1",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
]
